+++ src/stelnet/board/commodity/CommodityAction.java
@@ -8,12 +8,15 @@ import stelnet.board.commodity.market.SellMarketFactory;
 import stelnet.board.commodity.price.DemandPrice;
 import stelnet.board.commodity.price.Price;
 import stelnet.board.commodity.price.SupplyPrice;
+import stelnet.board.commodity.view.board.MultiMarketView;
+import stelnet.board.commodity.view.board.SingleMarketView;
+import uilib.RenderableFactory;
 
 public enum CommodityAction {
     BUY {
         @Override
         public List<MarketAPI> getMarkets(String commodityId) {
-            return new BuyMarketFactory(commodityId).createMarkets();
+            return getBuyMarkets(commodityId);
         }
 
         @Override
@@ -24,7 +27,7 @@ public enum CommodityAction {
     SELL {
         @Override
         public List<MarketAPI> getMarkets(String commodityId) {
-            return new SellMarketFactory(commodityId).createMarkets();
+            return getSellMarkets(commodityId);
         }
 
         @Override
@@ -35,13 +38,8 @@ public enum CommodityAction {
 
     PROFIT {
         @Override
-        public List<MarketAPI> getSellMarkets(String commodityId) {
-            return new SellMarketFactory(commodityId).createMarkets();
-        }
-
-        @Override
-        public List<MarketAPI> getBuyMarkets(String commodityId) {
-            return new BuyMarketFactory(commodityId).createMarkets();
+        public RenderableFactory getFactory(String commodityId, IntelTracker intelTracker) {
+            return new MultiMarketView(commodityId, this, intelTracker);
         }
 
         @Override
@@ -50,29 +48,23 @@ public enum CommodityAction {
         }
     };
 
+    public RenderableFactory getFactory(String commodityId, IntelTracker intelTracker) {
+        return new SingleMarketView(commodityId, this, intelTracker);
+    }
+
     public List<MarketAPI> getMarkets(String commodityId) {
         return Collections.emptyList();
     }
 
     public List<MarketAPI> getSellMarkets(String commodityId) {
-        return Collections.emptyList();
+        return new SellMarketFactory(commodityId).createMarkets();
     }
 
     public List<MarketAPI> getBuyMarkets(String commodityId) {
-        return Collections.emptyList();
+        return new BuyMarketFactory(commodityId).createMarkets();
     }
 
     public Price getPrice(String commodityId) {
-        return new Price() {
-            @Override
-            public float getPriceAmount(MarketAPI market) {
-                return 0;
-            }
-
-            @Override
-            public float getPriceAmount(MarketAPI market, int quantity) {
-                return 0;
-            }
-        };
+        return null;
     }
 }
