+++ src/stelnet/board/commodity/CommodityView.java
@@ -5,7 +5,6 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import stelnet.board.commodity.view.board.ButtonViewFactory;
 import stelnet.board.commodity.view.board.DeleteViewFactory;
-import stelnet.board.commodity.view.board.IntelViewFactory;
 import stelnet.board.commodity.view.board.TabViewFactory;
 import uilib.Renderable;
 import uilib.RenderableFactory;
@@ -22,7 +21,7 @@ public class CommodityView implements RenderableFactory {
     public List<Renderable> create(Size size) {
         List<Renderable> elements = new LinkedList<>();
         elements.addAll(new TabViewFactory(commodityId, activeTab).create(size));
-        elements.addAll(new IntelViewFactory(commodityId, activeTab, intelTracker).create(size));
+        elements.addAll(activeTab.getFactory(commodityId, intelTracker).create(size));
         elements.addAll(new ButtonViewFactory(commodityId).create(size));
         elements.addAll(new DeleteViewFactory(commodityId).create(size));
         return elements;
