+++ src/stelnet/board/commodity/view/board/SingleMarketButton.java
@@ -2,7 +2,6 @@ package stelnet.board.commodity.view.board;
 
 import com.fs.starfarer.api.campaign.econ.MarketAPI;
 import com.fs.starfarer.api.ui.IntelUIAPI;
-import stelnet.board.commodity.CommodityAction;
 import stelnet.board.commodity.IntelTracker;
 import uilib.AreaCheckbox;
 import uilib.EventHandler;
@@ -10,9 +9,9 @@ import uilib.property.Location;
 import uilib.property.Position;
 import uilib.property.Size;
 
-public class IntelButton extends AreaCheckbox {
+public class SingleMarketButton extends AreaCheckbox {
 
-    public IntelButton(int i, final String commodityId, final MarketAPI market, final IntelTracker tracker) {
+    public SingleMarketButton(int i, final String commodityId, final MarketAPI market, final IntelTracker tracker) {
         super(new Size(28f, 24f), String.valueOf(i), true, tracker.has(commodityId, market));
         setLocation(Location.BOTTOM_LEFT);
         setOffset(new Position(0, 1));
